Add Cilium security identity GC
The Cilium operator is also in charge of garbage collecting unused security identities. This ports the minimum code required to enable that same function within netreap. For simplicity, this also introduces a new "leader reaper" that wraps the functions that should only be executed by one of the netreap instances.