Fix endpoint reconcile handling of identifiers
The only identifier that is guaranteed to be on a Cilium endpoint is the CNI attachment ID. Previously the reconcile loop would look for the container ID but that can be empty string. The outcome was that the reconcile loop would skip over all endpoints rather than reconcile the labels.

Other parts of the code already used the CNI attachment ID correctly, so the reconcile loop now does too.

This change also updates the tests to use stretchr for mocks.